Critério 10 e ajustado para usar um header unico para area publica e area interna
