+++ src/components/lib/header/index.tsx
@@ -1,12 +1,16 @@
-import { Layout, Row, Space } from 'antd';
-import React from 'react';
+import { LoginOutlined } from '@ant-design/icons';
+import { Flex, Layout, Menu, MenuProps, Row, Space } from 'antd';
+import React, { useState } from 'react';
+import { Link, useLocation, useNavigate } from 'react-router-dom';
 import ConectaLogo from '~/assets/conecta-formacao-logo.svg';
+import { AreaPublicaEnum, MENU_AREA_PUBLICA } from '~/core/constants/menu-area-publica';
+import { ROUTES } from '~/core/enum/routes-enum';
+import { useAppSelector } from '~/core/hooks/use-redux';
 import { store } from '~/core/redux';
 import { setDeslogar } from '~/core/redux/modules/auth/actions';
 import { BoxShadow, Colors } from '~/core/styles/colors';
-import ExitButton from '../exit-button';
 import DropdownPerfil from '../dropdown-perfil';
-import { ROUTES } from '~/core/enum/routes-enum';
+import ExitButton from '../exit-button';
 
 const contentStyle: React.CSSProperties = {
   position: 'sticky',
@@ -20,21 +24,69 @@ const contentStyle: React.CSSProperties = {
   boxShadow: BoxShadow.DEFAULT,
 };
 
+const menuItems: MenuProps['items'] = [
+  {
+    label: <Link to={ROUTES.AREA_PUBLICA}>{MENU_AREA_PUBLICA.HOME.LABEL}</Link>,
+    key: MENU_AREA_PUBLICA.HOME.KEY,
+  },
+  {
+    label: <Link to={ROUTES.LOGIN}>{MENU_AREA_PUBLICA.LOGIN.LABEL}</Link>,
+    key: MENU_AREA_PUBLICA.LOGIN.KEY,
+    icon: <LoginOutlined />,
+  },
+];
+
 const Header: React.FC = () => {
+  const navigate = useNavigate();
+  const location = useLocation();
+  const [menu, setMenu] = useState(MENU_AREA_PUBLICA.HOME.LABEL);
+
+  const autenticado = useAppSelector((state) => state.auth.autenticado);
+
+  const rotaEhAreaPublica = location.pathname.includes(ROUTES.AREA_PUBLICA);
+
+  const filteredMenuItems = autenticado
+    ? menuItems.filter((item) => item?.key !== MENU_AREA_PUBLICA.LOGIN.KEY)
+    : menuItems;
+
+  const alterarMenu: MenuProps['onClick'] = (e) => {
+    setMenu(e.key);
+    if (e.key === AreaPublicaEnum.Home.toString()) {
+      irParaHome();
+    }
+  };
+
+  const irParaHome = () => navigate(ROUTES.AREA_PUBLICA, { replace: true });
+
   return (
     <Layout.Header style={contentStyle}>
-      <a href={ROUTES.PRINCIPAL}>
+      <a href={rotaEhAreaPublica ? ROUTES.AREA_PUBLICA : ROUTES.PRINCIPAL}>
         <img style={{ height: '50px' }} src={ConectaLogo} alt='Conecta Formação LOGO' />
       </a>
+
       <Row justify='end' style={{ width: '100%' }}>
-        <Space>
-          <DropdownPerfil />
-          <ExitButton
-            onClick={() => {
-              store.dispatch(setDeslogar());
-            }}
-          />
-        </Space>
+        <Flex gap={16}>
+          {rotaEhAreaPublica && (
+            <Menu
+              onClick={alterarMenu}
+              selectedKeys={[menu]}
+              mode='horizontal'
+              items={filteredMenuItems}
+              style={{ height: 70 }}
+            />
+          )}
+
+          {autenticado && (
+            <Space>
+              <DropdownPerfil />
+              <ExitButton
+                onClick={() => {
+                  store.dispatch(setDeslogar());
+                }}
+              />
+            </Space>
+          )}
+        </Flex>
       </Row>
     </Layout.Header>
   );
